Simplify input handling
* Move InputFrame in its own files
* Remove friendship of Pi on Input
