+++ src/KeyBindings.h
@@ -7,14 +7,7 @@
 #include "libs.h"
 #include <iosfwd>
 
-enum InputResponse {
-	// None of the inputs match the event.
-	RESPONSE_NOMATCH = 0,
-	// An input matched, but won't consume the event.
-	RESPONSE_PASSTHROUGH,
-	// An input matched and consumed the event.
-	RESPONSE_MATCHED
-};
+#include "InputFrame.h"
 
 namespace KeyBindings {
 	enum Type {
