+++ src/ship/ShipViewController.h
@@ -5,6 +5,7 @@
 
 #include "CameraController.h"
 #include "Input.h"
+#include "InputFrame.h"
 #include "InteractionController.h"
 #include "KeyBindings.h"
 #include "utils.h"
@@ -44,7 +45,7 @@ public:
 	// Here temporarely because of initialization order
 	void SetCamType(Ship *ship, enum CamType c);
 
-	static struct InputBinding : public Input::InputFrame {
+	static struct InputBinding : public InputFrame {
 		using Action = KeyBindings::ActionBinding;
 		using Axis = KeyBindings::AxisBinding;
